+++ wgpu/tests/resource_error.rs
@@ -44,6 +44,7 @@ fn bad_texture() {
                 dimension: wgpu::TextureDimension::D2,
                 format: wgpu::TextureFormat::Rgba8UnormSrgb,
                 usage: wgpu::TextureUsages::all(),
+                view_formats: &[],
             })
         });